refactor(core): relocate /percy/maestro-screenshot handler into maestro-screenshot.js
Move the ~440-line route handler body, plus the parsePngDimensions and
manualScreenshotWalk helpers, verbatim out of api.js into a dedicated
maestro-screenshot.js module exporting handleMaestroScreenshot(req, res, percy).
api.js now declares the route as a one-line delegate, mirroring the
handleComparisonUpload/handleSyncJob pattern. All istanbul-ignore annotations
travel verbatim with their statements. Move the file-level semgrep
path-traversal suppression from api.js to maestro-screenshot.js (the new owner
of the path.join sinks). No behavior change.

api.js drops from 994 to 383 lines.
