+++ .semgrepignore
@@ -20,17 +20,17 @@ packages/core/src/lock.js
 # every join. Suppress at the file level with this rationale.
 packages/core/src/archive.js
 
-# api.js path-traversal guards live at the top of the /percy/maestro-screenshot
-# route handler: `name` and `sessionId` are both validated against
-# /^[a-zA-Z0-9_-]+$/ (SAFE_ID, line ~322) BEFORE any path.join() runs.
+# maestro-screenshot.js path-traversal guards live at the top of the
+# handleMaestroScreenshot handler: `name` and `sessionId` are both validated
+# against /^[a-zA-Z0-9_-]+$/ (SAFE_ID) BEFORE any path.join() runs.
 # Traversal sequences (`..`, `/`, `\0`) are rejected with 400 there. The
-# fallback walker is also depth-capped at 15 levels. semgrep's
-# path-join-resolve-traversal rule cannot follow the regex validation
-# chain across the function body, so it flags the joins on lines 445
-# and 462. Inline `// nosemgrep` directives (preceding and same-line)
-# were not honored by the semgrep CI version in use — suppress at the
-# file level with this rationale.
-packages/core/src/api.js
+# fallback walker (manualScreenshotWalk) is also depth-capped at 15 levels.
+# semgrep's path-join-resolve-traversal rule cannot follow the regex
+# validation chain across the function body, so it flags the joins inside
+# manualScreenshotWalk. Inline `// nosemgrep` directives (preceding and
+# same-line) were not honored by the semgrep CI version in use — suppress at
+# the file level with this rationale.
+packages/core/src/maestro-screenshot.js
 
 # Test files load fixtures from hardcoded subpaths under
 # test/fixtures/maestro-{hierarchy,ios-hierarchy}. The fixture filenames
